adds authorization object to the context when executing a subscriber
this is needed because when the user is navigated to other screen there is, in
ConfigureUIServiceInitListener, a check if the the user has permission
to access the page. This is done by 'VaadinSecurityUtils.isAcessGranted',
which gets the Authentication object from the security context (
SecurityContextHolder.getContext()). The default strategy used for storing
the secure context is MODE_THREADLOCAL, i.e. store it in Thread Local.
However, the event bus subscribers are executed in custom threads from a pool
and, then, don't have access to Authorization.