1773: Update experiment nav bar when an experiment status changed draft/archived in other tab (#1786)
* 1773: update experiments nav bar when an experiment is archived or start training

* 1773: update experiments nav bar when an experiment starts training

* Add tests for started/archived experiment

* Fix unstable tests

* Fix failed test

* 1773: make sure experiments will be always sorted

* 1773: when a draft experiment starts training in other tab, redirect user to ExperimentView

* 1773: fix bug that redirected the user to the wrong page

* Add test Check experiment side bar when an experiment is created and stopped in other tab

* 1773: moved code that updates nav tree to the event handler

* 1773: refactor code to archive object to make sure we always pass the object

with the change we'll avoid one database read when archiving an experiment.

* adds authorization object to the context when executing a subscriber

this is needed because when the user is navigated to other screen there is, in
ConfigureUIServiceInitListener, a check if the the user has permission
to access the page. This is done by 'VaadinSecurityUtils.isAcessGranted',
which gets the Authentication object from the security context (
SecurityContextHolder.getContext()). The default strategy used for storing
the secure context is MODE_THREADLOCAL, i.e. store it in Thread Local.
However, the event bus subscribers are executed in custom threads from a pool
and, then, don't have access to Authorization.

Co-authored-by: Slin Lee <slin@pathmind.com>
Co-authored-by: EvgeniyEA <evegeniy@skymind.io>
Co-authored-by: EvgeniyEA <47972140+EvgeniyEA@users.noreply.github.com>
Co-authored-by: Alexander Makarov <830564+alexamakarov@users.noreply.github.com>